larry: Avoid shipping firmware
Starting ~ CPH246x_15.0.0.1600 firmware, any downgrades to
previous firmware bricks the device and puts the device in QDL mode.
Thus, its upto the user to be on any of the CPH246x_15.0.0.* firmwares

As a cautionary measure, any and all old builds are purged so that nobody
downloads and installs them.